refactor: rename route from `resend_activation` -> `reactivate`
