+++ src/routes/api/reactivate.ts
